pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "beast2analysisutils"
version = "0.1.0"
description = "Analysis utilities for BEAST2"
readme = "README.md"
authors = [
  { name = "Paula H. Weidemueller", email = "paula.weidemueller@gmail.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.7"
dependencies = [
    "numpy>=1.20",
    "pandas>=1.0",
    "biopython>=1.80",
    "dendropy>=4.5.2",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]
doc = [
    "mkdocs-material",
    "mkdocstrings[python]",
]

[tool.setuptools.packages.find]
where = ["src"]
